20250502_docker_bot_video/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
opencv-python==4.8.1.78
numpy==1.26.2
google-genai==1.13.0
